Refactor frame helpers to share more code (#500)
